refactor: extract Sportradar client types to separate file
Move SportradarConfig/SportradarClient/Sport/Tournament/Match/Team/Score/Odds/SportradarOddsResponse structs from client.go to types.go. Remove shopspring/decimal import from client.go.
